tests/wpt/tests/html/canvas/offscreen/compositing/2d.composite.image.lighter.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.composite.image.lighter</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.composite.image.lighter</h1>
<p class="desc"></p>


<script>
var t = async_test("");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  ctx.fillStyle = 'rgba(0, 255, 255, 0.5)';
  ctx.fillRect(0, 0, 100, 50);
  ctx.globalCompositeOperation = 'lighter';
  var promise = new Promise(function(resolve, reject) {
      var xhr = new XMLHttpRequest();
      xhr.open("GET", '/images/yellow75.png');
      xhr.responseType = 'blob';
      xhr.send();
      xhr.onload = function() {
          resolve(xhr.response);
      };
  });
  promise.then(function(response) {
      return createImageBitmap(response).then(bitmap => {
          ctx.drawImage(bitmap, 0, 0);
          _assertPixelApprox(canvas, 50,25, 191,255,128,255, 5);
      });
  }).then(t_pass, t_fail);

});
</script>
